Refactor LED circle direction and width calculations; move functions to locations module
