+++ Project/main/src/helperFunctions.c
@@ -224,97 +224,3 @@ void set_btnmatrix_led(uint8_t position){
 
 	btnmatrix_outSetMutexValue(data_button_matrix);
 }
-
-#define DIST_MAX_WIDTH 20	// Distance at which the circle has minimum width
-#define DIST_MIN_WIDTH 100 // Was 1000	//Distance at which the circle has maximum width
-#define DIST_RANGE (DIST_MIN_WIDTH - DIST_MAX_WIDTH)
-#define LEDS_MIN_WIDTH 7	// Amount of LEDS at DIST_MIN_WIDTH
-#define LEDS_MAX_WIDTH 32	// Amount of LEDS at DIST_MAX_WIDTH
-#define LEDS_RANGE_WIDTH (LEDS_MAX_WIDTH - LEDS_MIN_WIDTH)
-
-
-int get_led_width(int distance) {
-	if (distance > DIST_MIN_WIDTH) { return LEDS_MIN_WIDTH; }
-	if (distance < DIST_MAX_WIDTH) { return LEDS_MAX_WIDTH; }
-	distance -= DIST_MAX_WIDTH;
-	double distFraction = (double)distance / DIST_RANGE;
-	distFraction = 1 - distFraction;
-	return LEDS_RANGE_WIDTH*distFraction + LEDS_MIN_WIDTH;
-}
-
-int circleMovingAvg(int newValue) {
-	double rad = toRadians(newValue);
-	double sinVal = sin(rad);
-	double cosVal = cos(rad);
-
-	for (int i = ANGLE_BUFFER_SIZE-1; i > 0; i--) {	// Move every entry in the buffer up by one index
-		angleSinBuffer[i] = angleSinBuffer[i-1];
-		angleCosBuffer[i] = angleCosBuffer[i-1];
-	}
-	angleSinBuffer[0] = sinVal;						// Insert the new value
-	angleCosBuffer[0] = cosVal;
-
-	double sinSum = 0;
-	double cosSum = 0;
-	for (int i = 0; i < ANGLE_BUFFER_SIZE; i++) {
-		sinSum += angleSinBuffer[i];
-		cosSum += angleCosBuffer[i];
-	}
-	double sinAvg = sinSum / ANGLE_BUFFER_SIZE;
-	double cosAvg = cosSum / ANGLE_BUFFER_SIZE;
-	return toDegrees(atan2(sinAvg,cosAvg));
-}
-
-/**
- * @brief Function to set the led circle direction and distance for showing where the user should go next.
- * @param dir The direction in degrees (0-359) where the LEDs should be centered.
- * @param distance
- */
-void led_circle_dir_width(unsigned dir, int distance)
-{
-	const int nrPixels = 64;
-	const int maxDir = 359;
-	unsigned width = get_led_width(distance); // Convert distance to width in pixels
-
-	float centerFloat = (float)dir / maxDir;
-	int centerPixel = round(centerFloat * nrPixels);
-
-	int leftBound = centerPixel - (width / 2);
-	if (leftBound < 0 ) {
-		leftBound = nrPixels + leftBound;
-	}
-	int rightBound = centerPixel + (width / 2);
-	if (rightBound > (nrPixels - 1)) {
-		rightBound = rightBound - nrPixels;
-	}
-	bool overlap = false;
-	if (leftBound > rightBound) {
-		overlap = true;
-	}
-
-	const int nrBytes = 8;
-	const int nrBits = 8;
-	uint8_t outputValues[8] = {0,0,0,0,0,0,0,0};	// Initialize to zero (all LEDs off)
-	for (int byteCount = 0; byteCount < nrBytes; byteCount++) {
-		for (int bitCount = 0; bitCount < nrBits; bitCount++) {
-			int bitIndex = 8 * byteCount + bitCount;
-			//bool enableLed = (!overlap && (bitIndex >= leftBound) && (bitIndex >= rightBound)) || (overlap && (bitIndex >= leftBound || bitIndex <= rightBound));
-			bool enableLed = false;
-			if (!overlap) {
-				if (bitIndex >= leftBound && bitIndex <= rightBound) {
-					enableLed = true;
-				}
-			} else {
-				if (bitIndex >= leftBound || bitIndex <= rightBound) {
-					enableLed = true;
-				}
-			}
-			//LOG_WRN("Byte %d bit %d (bitindex %d): %d\n", byteCount, bitCount, bitIndex, enableLed);
-			if (enableLed) {
-				outputValues[byteCount] |= (1 << (7-bitCount));
-			}
-
-		}
-	}
-	ledcircleSetMutexValue(outputValues);
-}
